refactor: split aliases, composites and values for raw tokens
