+++ OUDS/Core/Tokens/RawTokens/Sources/Aliases/DimensionRawTokens+Aliases.swift
@@ -1,0 +1,17 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+
+/// Raw tokens of dimensions are finaly `Double`, to keep grammar clean and clear with design system grammar.
+public typealias DimensionRawToken = Double
