+++ OUDS/Core/Tokens/RawTokens/Sources/Aliases/TypographyRawTokens+Aliases.swift
@@ -1,0 +1,26 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import Foundation
+
+/// In the global design system tool, *font family* raw tokens are basically `String` values, to keep grammar clean and clear with design system grammar.
+public typealias TypographyFontFamilyRawToken = String
+
+/// In the global design system tool, *font weight* raw tokens are basically `String` values, to keep grammar clean and clear with design system grammar.
+public typealias TypographyFontWeightRawToken = String
+
+/// In the global design system tool, *font size* raw tokens are basically `CGFloat` values, to keep grammar clean and clear with design system grammar.
+public typealias TypographyFontSizeRawToken = CGFloat
+
+/// In the global design system tool, *font line height* raw tokens are basically `Int` values, to keep grammar clean and clear with design system grammar.
+public typealias TypographyFontLineHeightRawToken = CGFloat
