+++ OUDS/Core/Tokens/RawTokens/Sources/DimensionRawTokens.swift
@@ -13,15 +13,11 @@
 
 import Foundation
 
-// MARK: - Type aliases to keep grammar clear
-
-/// Raw tokens of dimensions are finaly `Double`
-public typealias DimensionRawToken = Double
-
-// MARK: - Raw tokens
-
-/// This is the group of all **raw tokens** related to **dimensions**.
+/// Defined as the group of all **raw tokens** related to **dimensions**.
 /// Primitive types such as `Double` must be used to as to allow to use `@objc` keywords in extensions for overriding.
 /// Such tokens are packed in a _Swift enum_ so as to gather them in one object with the suitable namespace and avoid to have just constants in nothing else
 ///  (i.e. publicly accessible from everywhere). More optimized than _struct_.
-public enum DimensionRawTokens { }
+public enum DimensionRawTokens {
+
+    // Note: So as to help the Figma JSON to Swift parser the values have been added in the Values folder
+}
